deleted excess files according to architecture, fixed some imports in main.py, gui.py, task_cards.py (без этих изменений не запускалось)
